Update script.js
Added and deleted irrelevant documentation. Added variables for computer score and computer result and the result for the end of each round. Attempting to make a variable in computerplay() that is the computerselection that can be used in the playround function to be used for the if comparisons for keeping track of the score