Update script.js
Added a while loop again, and it works this time. scoring system still doesnt work but it returns the result string which has the score variables. deleted the playerplay function that i added because i think i should be able to do that inside the playround function. for some reason the random index still work but isnt changed every round anymore.